+++ lib/rspec/matchers/built_in/include.rb
@@ -29,7 +29,9 @@ module RSpec
         def perform_match(predicate, hash_predicate, actuals, expecteds)
           expecteds.send(predicate) do |expected|
             if comparing_hash_values?(actuals, expected)
-              expected.send(hash_predicate) {|k,v| actuals[k] == v}
+              expected.send(hash_predicate) {|k,v|
+                actuals.has_key?(k) && actuals[k] == v
+              }
             elsif comparing_hash_keys?(actuals, expected)
               actuals.has_key?(expected)
             else