+++ Changelog.md
@@ -5,6 +5,11 @@ Enhancements
 
 * Colorize diffs if the `--color` option is configured. (Alex Coplan)
 
+Bug fixes
+
+* Fix `include` matcher so that `expect({}).to include(:a => nil)`
+  fails as it should (Sam Phippen).
+
 ### 2.11.3 / 2012-09-04
 [full changelog](http://github.com/rspec/rspec-expectations/compare/v2.11.2...v2.11.3)